+++ FirFilter.h
@@ -7,13 +7,13 @@
 class FirFilter
 {
 public:
-	FirFilter(const float32* coef, uint32 coeflen);
+	FirFilter(const double* coef, uint32 coeflen);
 	virtual ~FirFilter();
 	virtual void ProcessWav(WaveFile& _wav, uint32 _start = 0, int _len = 0x7fffffff);
 
 protected:
-	float32* h;		// 申请内存
-	float32* flt;	// 指向滤波器中心
+	double* h;		// 申请内存
+	double* flt;	// 指向滤波器中心
 	uint32 len;
 };
